Feat/tableabsensi

+++ resources/views/container/_modal.blade.php
@@ -85,6 +85,7 @@
   let modalBody = $('#crudModal').find('.modal-body').html()
 
   let dataMaxLength = []
+  var data_id
 
   $(document).ready(function() {
     $('#btnSubmit').click(function(event) {
@@ -199,6 +200,7 @@
     setFormBindKeys(form)
 
     activeGrid = null
+    data_id = $('#crudForm').find('[name=id]').val();
 
     form.find('#btnSubmit').prop('disabled', false)
     if (form.data('action') == "view") {
@@ -211,10 +213,41 @@
 
   $('#crudModal').on('hidden.bs.modal', () => {
     activeGrid = '#jqGrid'
+    removeEditingBy(data_id)
     $('#crudModal').find('.modal-body').html(modalBody)
   })
 
-  function cekValidasidelete(Id) {
+  function removeEditingBy(id) {
+    $.ajax({
+      url: `{{ config('app.api_url') }}bataledit`,
+      method: 'POST',
+      dataType: 'JSON',
+      headers: {
+        Authorization: `Bearer ${accessToken}`
+      },
+      data: {
+        id: id,
+        aksi: 'BATAL',
+        table: 'container'
+        
+      },
+      success: response => {
+        $("#crudModal").modal("hide")
+      },
+      error: error => {
+        if (error.status === 422) {
+          $('.is-invalid').removeClass('is-invalid')
+          $('.invalid-feedback').remove()
+          
+          setErrorMessages(form, error.responseJSON.errors);
+        } else {
+          showDialog(error.responseJSON)
+        }
+      },
+    })
+  }
+
+  function cekValidasi(Id,Aksi) {
     $.ajax({
       url: `{{ config('app.api_url') }}container/${Id}/cekValidasi`,
       method: 'POST',
@@ -223,7 +256,8 @@
         request.setRequestHeader('Authorization', `Bearer {{ session('access_token') }}`)
       },
       data: {
-        accessTokenTnl: accessTokenTnl
+        accessTokenTnl: accessTokenTnl,
+        aksi:Aksi
       },
       success: response => {
         var kondisi = response.kondisi
@@ -231,7 +265,11 @@
         if (kondisi == true) {
           showDialog(response.message['keterangan'])
         } else {
-          deleteContainer(Id)
+          if (Aksi=="edit") {
+            editContainer(Id)
+          }else if (Aksi=="delete"){
+            deleteContainer(Id)
+          }
         }
 
       }
